Add user admin and add new recette
I add an admin page who show the different user, that can permite to admin user to modify role or different value of user. I add also the adding part. There are a page to add new recette in database, so a page who are accesible only by chief, the page to see recette and a page to see one recette in particulary